requirements.txt
streamlit==0.69.0
Pillow
numpy
matplotlib
